requirements.txt
google-genai>=1.0
pypdfium2>=4.0
python-dotenv>=1.0
